requirements.txt
# Required pip installs
# pip install --force-reinstall -v "itsdangerous==2.0.1"
flask
flask-wtf
flask_sqlalchemy
email-validator
flask-bcrypt
flask-login
Pillow
itsdangerous == 2.0.1
flask-mail
